biosphere: Use Dessine-moi factories
