+++ eradiate/scenes/biosphere/__init__.py
@@ -1,11 +1,11 @@
-from ._core import BiosphereFactory, Canopy, CanopyElement, InstancedCanopyElement
+from ._core import biosphere_factory, Canopy, CanopyElement, InstancedCanopyElement
 from ._discrete import DiscreteCanopy
 from ._leaf_cloud import LeafCloud
 from ._tree import AbstractTree, MeshTree, MeshTreeElement
 
 __all__ = [
     "AbstractTree",
-    "BiosphereFactory",
+    "biosphere_factory",
     "Canopy",
     "CanopyElement",
     "DiscreteCanopy",
