+++ eradiate/scenes/biosphere/tests/test_biosphere_discrete.py
@@ -15,7 +15,9 @@ from eradiate.scenes.biosphere import (
 from eradiate.scenes.biosphere._discrete import DiscreteCanopy, LeafCloud
 from eradiate.scenes.core import KernelDict
 
-# -- Fixture definitions -------------------------------------------------------
+# ------------------------------------------------------------------------------
+#                            Fixture definitions
+# ------------------------------------------------------------------------------
 
 
 @pytest.fixture(scope="module")
@@ -96,99 +98,9 @@ f 5/6/6 1/12/6 8/11/6"""
         yield filename
 
 
-# -- InstancedCanopyElement tests ----------------------------------------------
-
-
-def test_instanced_canopy_element_create(mode_mono):
-    """Unit testing for the :class:`InstancedCanopyElement` constructor."""
-    cloud = LeafCloud(
-        leaf_positions=[[0, 0, 0]], leaf_orientations=[[0, 0, 1]], leaf_radii=[0.1]
-    )
-    tree = AbstractTree(
-        leaf_cloud=cloud, trunk_height=1.0, trunk_radius=0.1, trunk_reflectance=0.5
-    )
-    positions = np.array([[-5, -5, -5], [5, 5, 5]])
-    assert InstancedCanopyElement(canopy_element=cloud, instance_positions=positions)
-    assert InstancedCanopyElement(canopy_element=tree, instance_positions=positions)
-
-
-def test_instanced_canopy_element_kernel_dict(mode_mono):
-    """Unit testing for :meth:`InstancedLeafCloud.kernel_dict`."""
-    ctx = KernelDictContext()
-
-    cloud = LeafCloud(
-        leaf_positions=[[0, 0, 0]],
-        leaf_orientations=[[0, 0, 1]],
-        leaf_radii=[0.1],
-        id="leaf_cloud",
-    )
-    tree = AbstractTree(
-        leaf_cloud=cloud,
-        trunk_height=1.0,
-        trunk_radius=0.1,
-        trunk_reflectance=0.5,
-        id="tree",
-    )
-    positions = np.array([[-5, -5, -5], [5, 5, 5]])
-    kernel_dict = InstancedCanopyElement(
-        canopy_element=tree, instance_positions=positions, id="element"
-    ).kernel_dict(ctx=ctx)
-
-    # The generated kernel dictionary can be instantiated
-    assert KernelDict.new(kernel_dict).load()
-
-
-def test_instanced_leaf_cloud_from_file(mode_mono, tempfile_spheres):
-    """Unit testing for :meth:`InstancedLeafCloud.from_file`."""
-    assert InstancedCanopyElement.from_file(
-        filename=tempfile_spheres,
-        canopy_element=LeafCloud.from_dict(
-            {
-                "leaf_positions": [[0.0, 0.0, 0.0], [1.0, 1.0, 1.0]],
-                "leaf_orientations": [[0, 0, 1], [1, 0, 0]],
-                "leaf_radii": [0.1, 0.1],
-            }
-        ),
-    )
-
-
-def test_instanced_canopy_element_from_dict(mode_mono, tempfile_spheres):
-    """Unit testing for :meth:`InstancedLeafCloud.from_dict`."""
-    ctx = KernelDictContext()
-
-    # We can instantiate from a full-dict spec
-    instanced_leaf_cloud = InstancedCanopyElement.from_dict(
-        {
-            "canopy_element": {
-                "type": "leaf_cloud",
-                "leaf_positions": [[0, 0, 0], [1, 1, 1]],
-                "leaf_orientations": [[0, 0, 1], [1, 0, 0]],
-                "leaf_radii": [0.1, 0.1],
-            },
-            "instance_positions": np.array([[-5, -5, -5], [5, 5, 5]]),
-        }
-    )
-    assert instanced_leaf_cloud
-
-    # The generated kernel dictionary can be instantiated
-    assert KernelDict.new(instanced_leaf_cloud.kernel_dict(ctx=ctx)).load()
-
-    # We can access the from_file constructor from a dict
-    assert InstancedCanopyElement.from_dict(
-        {
-            "construct": "from_file",
-            "filename": tempfile_spheres,
-            "canopy_element": {
-                "type": "leaf_cloud",
-                "leaf_positions": [[0, 0, 0], [1, 1, 1]],
-                "leaf_orientations": [[0, 0, 1], [1, 0, 0]],
-                "leaf_radii": [0.1, 0.1],
-            },
-        }
-    )
-
-
-# -- DiscreteCanopy tests ------------------------------------------------------
+# ------------------------------------------------------------------------------
+#                             DiscreteCanopy Tests
+# ------------------------------------------------------------------------------
 
 
 def test_discrete_canopy_instantiate(mode_mono):
@@ -268,7 +180,7 @@ def test_discrete_canopy_advanced(
     A more advanced test where we load a series of different canopy elements:
 
     - A pre-computed canopy consisting of a generated cuboid leaf cloud
-    - A sereies of instanced leaf clouds from files
+    - A series of instanced leaf clouds from files
     - An abstract tree with a leaf cloud
     - A mesh based canopy element
