+++ eradiate/scenes/biosphere/tests/test_biosphere_tree_mesh.py
@@ -115,7 +115,7 @@ def tmpfile_stl():
 
 
 # ------------------------------------------------------------------------------
-#                                   Test
+#                                   Tests
 # ------------------------------------------------------------------------------
